// src/models/postnumbers.ts
export class Postnumbers {
  public _id: string;
  buisVoorHor: string;
  buisVoorVert: string;
  buisAchter: string;

  // nieuw

  buisVoorHorGres: string;
  buisVoorHorPVC: string;
  buisVoorHorPP: string;

  buisVoorVertGres: string;
  buisVoorVertPVC: string;
  buisVoorVertPP: string;

  bochtVoor: string;
  bochtAchter: string;
  bocht45Voor: string;
  bocht90Voor: string;
  bocht45Achter: string;
  bocht90Achter: string;
  reductieVoor: string;

  indrukmof: string;

  yStuk: string;
  tBuis: string;
  tStuk: string;
  flexAan: string;
  mof: string;
  krimpmof: string;
  koppelstuk: string;
  stop: string;
  andere: string;
  kunststof: string;
  pvcTStuk: string;
  gietIjzer: string;
  betonKader: string;
  aluKader: string;
  buisAchterGres: string;

  buisAchterPVC: string;
  buisAchterPP: string;

  YAchter: string;
  reductieAchter: string;
  funOmh: string;

  //RWA
  infiltratiePutje: string;
  aanslBovAfvoer: string;
  terugslagKlep: string;
  sifonPutje: string;
  tPutje: string;
  constructor() {}
}